+++ zingolib/src/testutils/send/transmit.rs
@@ -311,22 +311,13 @@ impl LightClient {
 
 #[cfg(test)]
 mod test {
-    //! all tests below (and in this mod) use example wallets, which describe real-world chains.
-
     use zingo_test_vectors::seeds;
 
     use crate::{
         config::{ClientConfig, WalletConfig},
-        lightclient::{LightClient, sync::test::sync_example_wallet},
+        lightclient::LightClient,
         mocks::proposal::ProposalBuilder,
-        testutils::{
-            chain_generics::{
-                conduct_chain::ConductChain as _, networked::NetworkedTestEnvironment,
-                with_assertions,
-            },
-            default_test_wallet_settings,
-        },
-        wallet::disk::testing::examples,
+        testutils::default_test_wallet_settings,
     };
 
     async fn create_basic_client() -> LightClient {
@@ -347,144 +338,5 @@ mod test {
         let proposal = ProposalBuilder::default().build();
         lc.send(proposal, zip32::AccountId::ZERO).await.unwrap_err();
         // TODO: match on specific error
-    }
-
-    /// live sync: execution time increases linearly until example wallet is upgraded
-    /// live send TESTNET: these assume the wallet has on-chain TAZ.
-    /// waits up to five blocks for confirmation per transaction. see [`zingolib/src/testutils/chain_generics/live_chain.rs`]
-    /// as of now, average block time is supposedly about 75 seconds
-    mod testnet {
-        use zcash_protocol::{PoolType, ShieldedPool};
-
-        use crate::testutils::lightclient::get_base_address;
-
-        use super::*;
-
-        #[ignore = "only one test can be run per testnet wallet at a time"]
-        #[tokio::test]
-        /// this is a networked sync test. its execution time scales linearly since last updated
-        /// this is a networked send test. whether it can work depends on the state of live wallet on the blockchain
-        async fn testnet_send_to_self_orchard_glory_goddess() {
-            let case =
-                examples::NetworkSeedVersion::Testnet(examples::TestnetSeedVersion::GloryGoddess);
-
-            let mut client = sync_example_wallet(case).await;
-
-            let client_addr =
-                get_base_address(&client, PoolType::Shielded(ShieldedPool::Orchard)).await;
-
-            with_assertions::assure_propose_send_bump_sync_all_recipients(
-                &mut NetworkedTestEnvironment::setup().await,
-                &mut client,
-                vec![(&client_addr, 20_000, None)],
-                vec![],
-                true,
-            )
-            .await
-            .unwrap();
-        }
-        #[ignore = "only one test can be run per testnet wallet at a time"]
-        #[tokio::test]
-        /// this is a networked sync test. its execution time scales linearly since last updated
-        /// this is a networked send test. whether it can work depends on the state of live wallet on the blockchain
-        async fn testnet_send_to_self_sapling_glory_goddess() {
-            let case =
-                examples::NetworkSeedVersion::Testnet(examples::TestnetSeedVersion::GloryGoddess);
-
-            let mut client = sync_example_wallet(case).await;
-
-            let client_addr =
-                get_base_address(&client, PoolType::Shielded(ShieldedPool::Sapling)).await;
-
-            with_assertions::assure_propose_send_bump_sync_all_recipients(
-                &mut NetworkedTestEnvironment::setup().await,
-                &mut client,
-                vec![(&client_addr, 20_000, None)],
-                vec![],
-                true,
-            )
-            .await
-            .unwrap();
-        }
-        #[ignore = "only one test can be run per testnet wallet at a time"]
-        #[tokio::test]
-        /// this is a networked sync test. its execution time scales linearly since last updated
-        /// this is a networked send test. whether it can work depends on the state of live wallet on the blockchain
-        /// about 273 seconds
-        async fn testnet_send_to_self_transparent_and_then_shield_glory_goddess() {
-            let case =
-                examples::NetworkSeedVersion::Testnet(examples::TestnetSeedVersion::GloryGoddess);
-
-            let mut client = sync_example_wallet(case).await;
-
-            let client_addr = get_base_address(&client, PoolType::Transparent).await;
-
-            let environment = &mut NetworkedTestEnvironment::setup().await;
-            with_assertions::assure_propose_send_bump_sync_all_recipients(
-                environment,
-                &mut client,
-                vec![(&client_addr, 100_001, None)],
-                vec![],
-                true,
-            )
-            .await
-            .unwrap();
-
-            let _ =
-                with_assertions::assure_propose_shield_bump_sync(environment, &mut client, true)
-                    .await
-                    .unwrap();
-        }
-        #[ignore = "this needs to pass CI, but we arent there with testnet"]
-        #[tokio::test]
-        /// this is a networked sync test. its execution time scales linearly since last updated
-        /// this is a networked send test. whether it can work depends on the state of live wallet on the blockchain
-        async fn testnet_send_to_self_all_pools_glory_goddess() {
-            let case =
-                examples::NetworkSeedVersion::Testnet(examples::TestnetSeedVersion::GloryGoddess);
-
-            let mut client = sync_example_wallet(case).await;
-            let environment = &mut NetworkedTestEnvironment::setup().await;
-
-            let client_addr =
-                get_base_address(&client, PoolType::Shielded(ShieldedPool::Orchard)).await;
-            with_assertions::assure_propose_send_bump_sync_all_recipients(
-                &mut NetworkedTestEnvironment::setup().await,
-                &mut client,
-                vec![(&client_addr, 14_000, None)],
-                vec![],
-                true,
-            )
-            .await
-            .unwrap();
-
-            let client_addr =
-                get_base_address(&client, PoolType::Shielded(ShieldedPool::Sapling)).await;
-            with_assertions::assure_propose_send_bump_sync_all_recipients(
-                &mut NetworkedTestEnvironment::setup().await,
-                &mut client,
-                vec![(&client_addr, 15_000, None)],
-                vec![],
-                true,
-            )
-            .await
-            .unwrap();
-
-            let client_addr = get_base_address(&client, PoolType::Transparent).await;
-            with_assertions::assure_propose_send_bump_sync_all_recipients(
-                environment,
-                &mut client,
-                vec![(&client_addr, 100_000, None)],
-                vec![],
-                true,
-            )
-            .await
-            .unwrap();
-
-            let _ =
-                with_assertions::assure_propose_shield_bump_sync(environment, &mut client, true)
-                    .await
-                    .unwrap();
-        }
     }
 }